Reorganize and clean code
