+++ client/src/shared/models/index.ts
@@ -11,6 +11,7 @@ export * from './FundingSource';
 export * from './FundingSourceTime';
 export * from './FundingTime';
 export * from './Gender';
+export * from './NewSite';
 export * from './Region';
 export * from './ObjectWithValidationErrors';
 export * from './UndefinableBoolean';
