+++ src/routes/organizations.ts
@@ -1,8 +1,5 @@
 import express, { Request, Response } from 'express';
-import { Organization, Site } from '../entity';
-import { getManager } from 'typeorm';
-import * as orgController from '../controllers/organizations';
-import * as siteController from '../controllers/sites';
+import * as controller from '../controllers/organizations';
 import { passAsyncError } from '../middleware/error/passAsyncError';
 import { ApiError, ForbiddenError, InternalServerError } from '../middleware/error/errors';
 
@@ -14,48 +11,8 @@ organizationsRouter.post(
     try {
       if (!req.user.isAdmin) throw new ForbiddenError();
       const { name, sites } = req.body;
-
-      // All or nothing approach to creating new orgs, so if either the
-      // org name is taken or *any* site name is taken, fail the whole thing
-      
-      // await controller.createOrganization(name);
-      
-      const orgNameExists = await orgController.doesOrgNameExist(name);
-      const siteNamesAllUnique = await siteController.areSiteNamesUnique(
-        sites.map((newSite) => newSite.name)
-      );
-      if (orgNameExists) {
-        throw new ResourceConflictError('Organization name already exists.');
-      }
-      else if (!siteNamesAllUnique) {
-        throw new ResourceConflictError('A site name is already taken.')
-      }
-      else {
-        
-        // Create org first so we can give its ID to created sites
-        const newOrg = await getManager().save(
-          getManager().create(Organization, { providerName: name })
-        );
-        const createdSites: Site[] = await Promise.all(sites.map(
-          async (s) => await getManager().save(
-            getManager().create(Site, {
-              siteName: s.name,
-              titleI: s.titleI,
-              region: s.region,
-              naeycId: s.naeycId,
-              registryId: s.registryId,
-              licenseNumber: parseInt(s.licenseNumber),
-              facilityCode: parseInt(s.facilityCode),
-              organization: newOrg,
-              organizationId: newOrg.id,
-            })
-          )
-        ));        
-        newOrg.sites = createdSites;
-        await getManager().save(newOrg);
-        res.sendStatus(201);
-      }
-
+      await controller.createOrganization(name, sites);
+      res.sendStatus(201);
     } catch (err) {
       if (err instanceof ApiError) throw err;
       console.error('Error creating organization: ', err);
